refactor: components split
